Ability to fetch value as int.

+++ src/main/java/com/fluidbpm/fluidwebkit/backing/bean/workspace/form/WebKitOpenFormConversationBean.java
@@ -251,9 +251,8 @@ public class WebKitOpenFormConversationBean extends ABaseManagedBean {
 			FlowItemClient flowItemClient = this.getFluidClientDS().getFlowItemClient();
 			try {
 				Long fluidItemFormId = wfiParam.getFluidItemFormId();
-				if (fluidItemFormId != null && fluidItemFormId.longValue() > 0) {
-					fluidItem = flowItemClient.getFluidItemByFormId(
-							fluidItemFormId, true, webKitForm.isEnableCalculatedLabels());
+				if (fluidItemFormId != null && fluidItemFormId > 0) {
+					fluidItem = flowItemClient.getFluidItemByFormId(fluidItemFormId, true, webKitForm.isEnableCalculatedLabels());
 				}
 			} catch (FluidClientException fce) {
 				if (FluidClientException.ErrorCode.NO_RESULT != fce.getErrorCode()) throw fce;
